services/orders-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic[email]==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
structlog==24.4.0
redis[hiredis]==5.2.1
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
aiosqlite==0.20.0
fakeredis[aioredis]==2.26.2
